refactor: finish deferred cleanup — split UI partials, extract GpuFloor, primary ctors
- SpectreBootstrapUserInterface split into 5 partial files (Theme, Picker,
  GpuWarning, Progress + main) with a new IsInteractiveTerminal() helper that
  consolidates the duplicate redirected-stdin/interactive-capability check.
- Extract GpuFloor with MinDriver / MinCompute / CompareVersion / TryParseVersion
  so the UI and tests can read the floors without pulling in the probe services.
  NvidiaGpuPreflightCheck keeps re-exports of MinDriver/MinCompute for callers
  that already referenced them.
- Lock gate in UserContentWatcher uses System.Threading.Lock (.NET 9) for a
  cheaper hot-path and compiler-enforced scoped usage.
- Primary constructors for AssetDownloader, AssetPlanner, NvidiaSmiRunner,
  NvidiaRedistClient, HuggingFaceClient, PlanItemAssetDownloader,
  InstallStateLockStore where the ctor was plain parameter-to-field wiring.
- ArgumentNullException.ThrowIfNull on the IBootstrapUserInterface public
  surface (picker, GPU warning, progress, result).

Not addressed, with rationale:
- JsonSerializerContext source-gen for ManifestLoader: the custom
  AssetSourceJsonConverter doubles the `type` field as both polymorphic
  discriminator and the record's Type property, which STJ's [JsonPolymorphic]
  does not support. Migrating would require a wire-format change that could
  invalidate existing install-state.lock.json files. The app is not trimmed or
  AOT-published, so the reflection cost is negligible (~5 ms at startup).
- Shared PartialFileWriter: AssetDownloader (HTTP stream with resume + SHA256
  verify + optional PostProcess) and ZipExtractor (per-entry copy loop) differ
  enough that a shared abstraction would be either leaky or trivial.

All 571 tests (113 bootstrapper + 458 lib) pass. CSharpier clean.
