+++ src/PersonaEngine/PersonaEngine.Lib.Bootstrapper/Sources/NvidiaRedistClient.cs
@@ -4,25 +4,18 @@ using PersonaEngine.Lib.Assets.Manifest;
 
 namespace PersonaEngine.Lib.Bootstrapper.Sources;
 
-public sealed class NvidiaRedistClient : IAssetSource
+public sealed class NvidiaRedistClient(HttpClient http, ILogger<NvidiaRedistClient> logger)
+    : IAssetSource
 {
     public const string CudaArchiveBaseUrl =
         "https://developer.download.nvidia.com/compute/cuda/redist/";
     public const string CudnnArchiveBaseUrl =
         "https://developer.download.nvidia.com/compute/cudnn/redist/";
 
-    private readonly HttpClient _http;
-    private readonly ILogger<NvidiaRedistClient> _logger;
     private readonly ConcurrentDictionary<string, Task<NvidiaRedistManifest>> _manifestCache =
         new();
 
     public SourceType Type => SourceType.NvidiaRedist;
-
-    public NvidiaRedistClient(HttpClient http, ILogger<NvidiaRedistClient> logger)
-    {
-        _http = http;
-        _logger = logger;
-    }
 
     public async Task<AssetDownload> ResolveAsync(
         AssetEntry asset,
@@ -70,8 +63,8 @@ public sealed class NvidiaRedistClient : IAssetSource
 
     private async Task<NvidiaRedistManifest> FetchManifestAsync(string url, CancellationToken ct)
     {
-        _logger.LogDebug("Fetching NVIDIA redist manifest {Url}", url);
-        using var resp = await _http.GetAsync(url, ct).ConfigureAwait(false);
+        logger.LogDebug("Fetching NVIDIA redist manifest {Url}", url);
+        using var resp = await http.GetAsync(url, ct).ConfigureAwait(false);
         resp.EnsureSuccessStatusCode();
         var json = await resp.Content.ReadAsStringAsync(ct).ConfigureAwait(false);
         return NvidiaRedistManifest.Parse(json);
