+++ src/PersonaEngine/PersonaEngine.Lib.Bootstrapper/Sources/PlanItemAssetDownloader.cs
@@ -8,45 +8,33 @@ namespace PersonaEngine.Lib.Bootstrapper.Sources;
 /// Bridges <see cref="IAssetDownloader"/> to the underlying <see cref="AssetDownloader"/> by
 /// resolving the source URL via the correct source client and computing the destination path.
 /// </summary>
-public sealed class PlanItemAssetDownloader : IAssetDownloader
+public sealed class PlanItemAssetDownloader(
+    AssetDownloader inner,
+    HuggingFaceClient hf,
+    NvidiaRedistClient nv,
+    string resourceRoot,
+    ILogger<PlanItemAssetDownloader>? log = null
+) : IAssetDownloader
 {
-    private readonly AssetDownloader _inner;
-    private readonly HuggingFaceClient _hf;
-    private readonly NvidiaRedistClient _nv;
-    private readonly string _resourceRoot;
-    private readonly ILogger<PlanItemAssetDownloader>? _log;
-
-    public PlanItemAssetDownloader(
-        AssetDownloader inner,
-        HuggingFaceClient hf,
-        NvidiaRedistClient nv,
-        string resourceRoot,
-        ILogger<PlanItemAssetDownloader>? log = null
-    )
-    {
-        _inner = inner;
-        _hf = hf;
-        _nv = nv;
-        _resourceRoot = resourceRoot;
-        _log = log;
-    }
-
     public async Task DownloadAsync(
         AssetPlanItem item,
         IProgress<DownloadProgress> progress,
         CancellationToken ct
     )
     {
+        ArgumentNullException.ThrowIfNull(item);
+        ArgumentNullException.ThrowIfNull(progress);
+
         IAssetSource client = item.Entry.Source.Type switch
         {
-            SourceType.HuggingFace => _hf,
-            SourceType.NvidiaRedist => _nv,
+            SourceType.HuggingFace => hf,
+            SourceType.NvidiaRedist => nv,
             _ => throw new InvalidOperationException(
                 $"Unknown source type '{item.Entry.Source.Type}' for asset '{item.Entry.Id}'"
             ),
         };
 
-        _log?.LogDebug(
+        log?.LogDebug(
             "Resolving download for asset '{AssetId}' via {Source}",
             item.Entry.Id,
             item.Entry.Source.Type
@@ -56,11 +44,11 @@ public sealed class PlanItemAssetDownloader : IAssetDownloader
         // to both the source client (for extraction targets) and the inner
         // AssetDownloader (for non-archive file moves). This is the single
         // source of truth for where the asset lands on disk.
-        var destinationPath = Path.Combine(_resourceRoot, item.Entry.InstallPath);
+        var destinationPath = Path.Combine(resourceRoot, item.Entry.InstallPath);
         var download = await client
             .ResolveAsync(item.Entry, destinationPath, ct)
             .ConfigureAwait(false);
 
-        await _inner.DownloadAsync(download, destinationPath, progress, ct).ConfigureAwait(false);
+        await inner.DownloadAsync(download, destinationPath, progress, ct).ConfigureAwait(false);
     }
 }
